webapp/requirements/dev.txt
-r base.txt

django-debug-toolbar==1.4
django-extensions==1.6.7